drop offlineFulfillmentFeatures
