+++ apps/fulfillment/src/app.ts
@@ -1,5 +1,5 @@
 import { appBuilder } from '@spryker-oryx/application';
-import { offlineFulfillmentFeatures } from '@spryker-oryx/presets/fulfillment';
+import { fulfillmentFeatures } from '@spryker-oryx/presets/fulfillment';
 import { fulfillmentTheme } from '@spryker-oryx/themes';
 import { ColorMode } from '@spryker-oryx/utilities';
 
@@ -8,7 +8,7 @@ const env = import.meta.env;
 appBuilder()
   .withEnvironment(env)
   .withFeature(
-    offlineFulfillmentFeatures({
+    fulfillmentFeatures({
       picking: { appVersion: env.ORYX_FULFILLMENT_APP_VERSION },
     })
   )
