+++ libs/template/presets/fulfillment/app.ts
@@ -5,7 +5,13 @@ import { AppFeature, coreFeature } from '@spryker-oryx/core';
 import { Resources, experienceFeature } from '@spryker-oryx/experience';
 import { formFeature } from '@spryker-oryx/form';
 import { I18nFeature, I18nFeatureOptions } from '@spryker-oryx/i18n';
+import {
+  IndexedDbFeature,
+  IndexedDbFeatureConfig,
+} from '@spryker-oryx/indexed-db';
+import { OfflineFeature } from '@spryker-oryx/offline';
 import { PickingFeature, PickingFeatureConfig } from '@spryker-oryx/picking';
+import { OfflinePickingFeature } from '@spryker-oryx/picking/offline';
 import { WebPushNotificationFeature } from '@spryker-oryx/push-notification/web';
 import {
   commonGraphics,
@@ -27,7 +33,7 @@ delete applicationFeature.plugins;
 export function fulfillmentFeatures(
   config?: FulfillmentFeaturesConfig
 ): AppFeature[] {
-  return [
+  const onlineFeatures = [
     uiFeature,
     coreFeature,
     ...(featureVersion >= '1.2'
@@ -46,6 +52,21 @@ export function fulfillmentFeatures(
     new PickingFeature(config?.picking),
     StaticExperienceFeature,
   ];
+
+  if (featureVersion >= '1.3') {
+    config = {
+      ...defaultFulfillmentConfig,
+      ...config,
+    };
+    return [
+      ...onlineFeatures,
+      new IndexedDbFeature(config?.indexedDb),
+      new OfflineFeature(),
+      new OfflinePickingFeature(),
+    ];
+  } else {
+    return onlineFeatures;
+  }
 }
 
 export interface FulfillmentFeaturesConfig {
@@ -55,9 +76,58 @@ export interface FulfillmentFeaturesConfig {
   fulfillmentRoot?: FulfillmentRootFeatureConfig;
   picking?: PickingFeatureConfig;
   i18n?: I18nFeatureOptions;
+
+  /**
+   * @since version 1.3
+   */
+  indexedDb?: IndexedDbFeatureConfig;
 }
 
 export const fulfillmentResources: Resources = {
   graphics: { ...commonGraphics, ...fulfillmentResourceGraphics },
   fonts: materialDesignLink,
 };
+
+/**
+ * @deprecated Since version 1.3, indexedDb is now part of the standard fulfillmentFeatures config.
+ */
+export interface SharedOfflineFulfillmentFeaturesConfig {
+  indexedDb?: IndexedDbFeatureConfig;
+}
+
+/**
+ * @deprecated Since version 1.3, use FulfillmentFeaturesConfig instead.
+ */
+export interface OfflineFulfillmentFeaturesConfig
+  extends FulfillmentFeaturesConfig,
+    SharedOfflineFulfillmentFeaturesConfig {}
+
+/**
+ * @deprecated Since version 1.3, use defaultFulfillmentConfig instead.
+ */
+export const defaultOfflineFulfillmentConfig: SharedOfflineFulfillmentFeaturesConfig =
+  {
+    indexedDb: { dbName: 'fulfillment-app-db' },
+  };
+
+export const defaultFulfillmentConfig: FulfillmentFeaturesConfig =
+  defaultOfflineFulfillmentConfig;
+
+/**
+ * @deprecated Since version 1.3, use fulfillmentFeatures instead.
+ */
+export function offlineFulfillmentFeatures(
+  config?: OfflineFulfillmentFeaturesConfig
+): AppFeature[] {
+  config = {
+    ...defaultOfflineFulfillmentConfig,
+    ...config,
+  };
+
+  return [
+    ...fulfillmentFeatures(config),
+    new IndexedDbFeature(config?.indexedDb),
+    new OfflineFeature(),
+    new OfflinePickingFeature(),
+  ];
+}
