+++ libs/template/presets/fulfillment/index.ts
@@ -1,4 +1,3 @@
 export * from './app';
 export * from './components';
 export * from './feature';
-export * from './offline';
